Intercept UseApplicationFramework property.
I ended up merging the two interceptor classes in order to achieve this.
